# src/configs/apidocs/api_docs_v1.yaml
openapi: 3.0.1
info:
  title: Human Resource Management System
  contact:
    email: dothienbinh13@gmail.com
  version: 1.0.0
  servers:
    - url: http://localhost:3000/
  tags:
    - name: users
      description: Everythong about users
    - name: jobs
      description: Everythong about job form
    - name: reviews
      description: Everythong about review form
paths:
...
  /job/{id}/updateFormJob:
    put:
      tags:
        - FormJob
      summary: user update FormJob
      description: user update FormJob
      operationId: updateFormJob
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              allOf:
                - $ref: "#/components/schemas/FormJob"
              required:
                - Name
                - Desc
                - Status
                - StatusDesc
              example:
                Name: thu viec 30
                Desc: abcabc abc abc
                Status: 0
                StatusDesc: chu hoan thanh
      responses:
        200:
          description: "ok"
          content:
            application/json:
              example:
                Message: Save success!!!
        400:
          content:
            application/json:
              example:
                errCode: 1
                status: "fail"
                typeError: Token Error
                statusError: Bad request
                message: err verify jwterr verify jwt
        500:
          description: Database Error
          content:
            application/json:
              example:
                errCode: 2
                status: "fail"
                typeError: Database Error
                statusError: Internal Server Error
                message: update fail !!! or formjob not found !!!
  /job/{id}/deleteFormJob:
    delete:
      tags:
        - FormJob
      summary: user update FormJob
      description: user update FormJob
      operationId: updateFormJob
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        200:
          description: "ok"
          content:
            application/json:
              example:
                Message: Delete success!!!
        400:
          content:
            application/json:
              example:
                errCode: 1
                status: "fail"
                typeError: Token Error
                statusError: Bad request
                message: err verify jwterr verify jwt
        500:
          description: Database Error
          content:
            application/json:
              example:
                errCode: 2
                status: "fail"
                typeError: Database Error
                statusError: Internal Server Error
                message: Delete fail !!! or formjob not found !!!
...
  /review/{id}/deleteFormReview:
    delete:
      tags:
        - FormReview
      summary: user delete FormReview
      description: user delete FormReview
      operationId: updateFormReview
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        200:
          description: "ok"
          content:
            application/json:
              example:
                Message: Delete success!!!
        400:
          content:
            application/json:
              example:
                errCode: 1
                status: "fail"
                typeError: Token Error
                statusError: Bad request
                message: err verify jwterr verify jwt
        500:
          description: Database Error
          content:
            application/json:
              example:
                errCode: 2
                status: "fail"
                typeError: Database Error
                statusError: Internal Server Error
                message: Delete fail !!!
components:
  schemas:
...
    FormJob:
      title: FormJob
      type: object
      properties:
        Name:
          type: string
        Desc:
          type: string
        Status:
          type: integer
        StatusDesc:
          type: string
...
  securitySchemes:
    ACCESS_TOKEN: # arbitrary name for the security scheme
      type: http
      scheme: basic
      bearerFormat: JWT
      in: cookie